requirements.txt
discordpy
pyautogui
pillow
python-dotenv
